pyproject.toml
[project]
name = "v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-htmx>=1.27.0",
    "djangorestframework>=3.16.1",
    "google-generativeai>=0.8.5",
    "ipdb>=0.13.13",
    "openai>=2.9.0",
]
